Exclusão das anotações

+++ SimpleNotes/SimpleNotes/src/main/java/api/example/SimpleNotes/domain/note/NoteRepository.java
@@ -5,10 +5,8 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.List;
-
 @Repository
-public interface NoteRepository extends JpaRepository<Note, Integer> {
+public interface NoteRepository extends JpaRepository<Note, Long> {
 
     Page<Note> findAllByWalletId(Long walletId, Pageable pageable);
 }
